requirements.txt
termcolor
python-dotenv
pfruck-contabo